+++ lib/AuthRocket.php
@@ -29,7 +29,7 @@ class RecordNotFound extends Error {}
 
 class AuthRocket {
 
-  const VERSION = '1.2.1';
+  const VERSION = '1.3.0';
 
 
   private $api;