+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### 1.3.0
 
 - Use system's CA certificate store instead
+- Fix parsing error
 
 #### 1.2.1